# swagger.yaml
openapi: 3.0.0
info:
  title: kubuka kwesu tobecisi cabutonga
  description: Kubuka is an organisation
  version: 1.0.0
servers:
  - url: https://server.kubukaspace.com
    description: Online development server
paths:
...
  /bookings/create:
    post:
      summary: Book for your venue and meals
      description: The booking office is not responsible for loss of materials during events
      operationId: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_names:
                  type: string
                  example: "Author Agent"
                company_name:
                  type: string
                  example: "Getme"
                company_address:
                  type: string
                  example: "Male 2355 mkhb3"
                agent_name:
                  type: string
                  example: "Mape Gweru"
                agent_phone:
                  type: string
                  example: "04411552320"
                numberOfDays:
                  type: string
                  example: "5"
                agent_email:
                  type: string
                  example: "user01@kubuka.com"
                numberOfGuests:
                  type: string
                  example: "23"
                from_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                to_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                venue_type:
                  type: string
                  example: "Boord room"
                price:
                  type: string
                  example: "202"
              required:
                - full_names
                - company_name
                - company_address
                - agent_name
                - agent_email
                - agent_phone
                - numberOfDays
                - numberOfGuests
                - from_date
                - to_date
                - venue_type
                - price
      responses:
        "200":
          description: The Cart made to the organisation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cart successfully made"
  /bookings/{id}:
    parameters:
      - in: path
        name: id
        description: Booking id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single booking by id
      description: Admin users can access any booking, regular users are restricted to their own booking.
      operationId: getBookingById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified booking
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  full_names:
                    type: string
                    example: "Author Minuusd"
                  company_name:
                    type: string
                    example: "Getme"
                  company_address:
                    type: string
                    example: "45 manho gerus"
                  agent_name:
                    type: string
                    example: "minuhed Mig"
                  agent_phone:
                    type: string
                    example: "0412251745"
                  numberOfDays:
                    type: string
                    example: "5"
                  agent_email:
                    type: string
                    example: "user01@kubuka.com"
                  numberOfGuests:
                    type: string
                    example: "4"
                  venue_type:
                    type: string
                    example: "Igzin for"
                  price:
                    type: string
                    example: "2020"
                  from_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  to_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete an booking
      description: Admin users can delete any booking, regular users are restricted to their own booking.
      operationId: deleteBooking
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Booking deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Booking deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /blog:
    post:
      summary: Make your nice post
      description: Save a blog
      operationId: createBlog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  exaple: "Kubuka hubitual"
                email:
                  type: string
                  example: "info@kubuka.com"
                phone:
                  type: string
                  example: "0994455660"
                blog_title:
                  type: string
                  example: "Information for bakwesu"
                blog_content:
                  type: string
                  example: "Thanks kubuka, I would like to invest."
              required:
                - name
                - email
                - phone
                - blog_title
                - blog_content
      responses:
        "200":
          description: Blog post has been created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Blog post created"
    get:
      summary: Get a list of all blogs
      description: All visitors and users can see blogs.
      operationId: getAllBlogs
      responses:
        "200":
          description: An array of all blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    name:
                      type: string
                      exaple: "Kubuka hubitual"
                    email:
                      type: string
                      example: "info@kubuka.com"
                    phone:
                      type: string
                      example: "0994455660"
                    blog_title:
                      type: string
                      example: "Information for bakwesu"
                    blog_content:
                      type: string
                      example: "Thanks kubuka, I would like to invest."
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /blog/{id}:
    parameters:
      - in: path
        name: id
        description: Blog id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single blog by id
      description: Admin users can access any blog, regular users are restricted to their own blog.
      operationId: getBlogById
      responses:
        "200":
          description: Details of the blog
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  name:
                    type: string
                    exaple: "Kubuka hubitual"
                  email:
                    type: string
                    example: "info@kubuka.com"
                  phone:
                    type: string
                    example: "0994455660"
                  blog_title:
                    type: string
                    example: "Information for bakwesu"
                  blog_content:
                    type: string
                    example: "Thanks kubuka, I would like to invest."
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
    put:
      summary: Update a blog
      description: Owner of the blog can update, admins can also update certin parts only
      operationId: updateBlog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  exaple: "Kubuka hubitual"
                email:
                  type: string
                  example: "info@kubuka.com"
                phone:
                  type: string
                  example: "0994455660"
                blog_title:
                  type: string
                  example: "Information for bakwesu"
                blog_content:
                  type: string
                  example: "Thanks kubuka, I would like to invest."
      responses:
        "200":
          description: Blog updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  name:
                    type: string
                    exaple: "Kubuka hubitual"
                  email:
                    type: string
                    example: "info@kubuka.com"
                  phone:
                    type: string
                    example: "0994455660"
                  blog_title:
                    type: string
                    example: "Information for bakwesu"
                  blog_content:
                    type: string
                    example: "Thanks kubuka, I would like to invest."
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete a Blog
      description: Admin users can delete any Blog, regular users are restricted to their own Blog.
      operationId: deleteCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blog deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Blog deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"